--- buffer-memory-segment/.dockerignore
@@ -1,6 +1,0 @@
-.DS_Store
-*.iml
-target
-.idea
-*.jfr
-.git
