--- buffer-memory-segment/src/main/resources/META-INF/services/io.netty5.buffer.MemoryManager
@@ -1,1 +1,0 @@
-io.netty5.buffer.memseg.SegmentMemoryManager
