+++ buffer/src/main/java/io/netty5/buffer/memseg/package-info.java
