+++ buffer/src/main/resources/META-INF/services/io.netty5.buffer.MemoryManager
@@ -1,2 +1,3 @@
 io.netty5.buffer.bytebuffer.ByteBufferMemoryManager
 io.netty5.buffer.unsafe.UnsafeMemoryManager
+io.netty5.buffer.memseg.SegmentMemoryManager
